requirements.txt
selenium
python-dotenv
google-generativeai
Flask
bs4